% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds Ratios and 95\% Confidence Intervals}
\usage{
OR_95CI(coef, se, siglevel = 0.05, roundto = 3)
}
\arguments{
\item{coef}{Numeric vector of coefficient estimates}

\item{se}{Numeric vector of standard errors}

\item{siglevel}{Significance level for the confidence interval}

\item{roundto}{Number of decimal places to round to}
}
\value{
A character vector of odds ratios and their 95\% confidence intervals.
}
\description{
This function calculates the odds ratios and 95\% confidence intervals (CI) from vectors of coefficients and standard errors.
}
\examples{
OR_95CI(c(0.5,-0.3),c(0.1,0.2))
}
\author{
Victoria Truong
}
