% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy Data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{y}{Binary outcome variable (0 or 1)}
\item{x1}{Continuous predictor}
\item{x2}{Binary predictor (0 or 1)}
}
}
\usage{
toydata
}
\description{
This is a dataset that contains a binary outcome and two predictors.
It is used to show logistic regression and the calculation of the
odds ratios and confidence intervals.
}
\keyword{datasets}
